+++ mipac/manager/admins/__init__.py
@@ -1,3 +1,4 @@
+from .ad import *
 from .admin import *
 from .announcement import *
 from .emoji import *