+++ mipac/models/channel.py
@@ -2,57 +2,17 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from mipac.models.lite.channel import ChannelLite
 from mipac.types.channel import IChannel
 
 if TYPE_CHECKING:
     from mipac.manager import ClientManager
 
 
-class Channel:
+class Channel(ChannelLite[IChannel]):
     def __init__(self, channel: IChannel, *, client: ClientManager) -> None:
-        self.__channel: IChannel = channel
-        self.__client: ClientManager = client
-
-    @property
-    def id(self) -> str:
-        return self.__channel['id']
-
-    @property
-    def created_at(self) -> str:
-        return self.__channel['created_at']
-
-    @property
-    def last_noted_at(self) -> str:
-        return self.__channel['last_noted_at']
-
-    @property
-    def name(self) -> str:
-        return self.__channel['name']
-
-    @property
-    def description(self) -> str | None:
-        return self.__channel['description']
-
-    @property
-    def banner_url(self) -> str | None:
-        return self.__channel['banner_url']
-
-    @property
-    def notes_count(self) -> int:
-        return self.__channel['notes_count']
-
-    @property
-    def users_count(self) -> int:
-        return self.__channel['users_count']
-
-    @property
-    def is_following(self) -> bool:
-        return bool(self.__channel['is_following'])
-
-    @property
-    def user_id(self) -> str:
-        return self.__channel['user_id']
+        super().__init__(channel=channel, client=client)
 
     @property
     def has_unread_note(self) -> bool:
-        return self.__channel['has_unread_note']
+        return self._channel['has_unread_note']