+++ mipac/types/channel.py
@@ -1,10 +1,10 @@
 from typing import TypedDict
 
 
-class IChannel(TypedDict):
+class IChannelLite(TypedDict):
     id: str
     created_at: str
-    last_noted_at: str
+    last_noted_at: str | None
     name: str
     description: str | None
     banner_url: str | None
@@ -12,4 +12,6 @@ class IChannel(TypedDict):
     users_count: int
     is_following: bool
     user_id: str
+
+class IChannel(IChannelLite):
     has_unread_note: bool